[Refactor Startup] fix wallet creation and restoring (#2392)
* Start create wallet machine when creating wallet

* Go to create wallet machine when creating wallet

* clean up

* Go to confirm seed

* some fixes

* Creating wallet

* Set selected wallet when creating a wallet

* Fix bugs when creating wallet

* Start restore wallet

* Move logic to state machine

* changes at existing seed

* Missing file

* Restore wallet is working

* Fix Error not showing

* lint fix

* fix duplicated translation ids

+++ app/main_dev/ipc.js
@@ -100,12 +100,14 @@ export const createWallet = (testnet, walletPath) => {
 };
 
 export const removeWallet = (testnet, walletPath) => {
+  if (!walletPath) return;
   let removeWalletDirectory = getWalletPath(testnet, walletPath);
   try {
     if (fs.pathExistsSync(removeWalletDirectory)) {
       fs.removeSync(removeWalletDirectory);
+      return true;
     }
-    return true;
+    return false;
   } catch (e) {
     logger.log("error", "error creating wallet: " + e);
     return false;
